+++ src/main/kotlin/io/openfuture/chain/controller/CryptoController.kt
@@ -7,6 +7,7 @@ import io.openfuture.chain.domain.crypto.key.KeyDto
 import io.openfuture.chain.service.CryptoService
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 
@@ -36,5 +37,8 @@ class CryptoController(
                 key.ecKey.getAddress()
         )
     }
+
+    @GetMapping("/generate")
+    fun generateKey() = cryptoService.generateKey()
 
 }