+++ src/main/kotlin/io/openfuture/chain/service/DefaultCryptoService.kt
@@ -7,6 +7,8 @@ import io.openfuture.chain.crypto.seed.PhraseLength.TWELVE
 import io.openfuture.chain.crypto.seed.calculator.SeedCalculator
 import io.openfuture.chain.crypto.seed.generator.SeedPhraseGenerator
 import io.openfuture.chain.crypto.seed.validator.SeedPhraseValidator
+import io.openfuture.chain.domain.crypto.key.AddressKeyDto
+import io.openfuture.chain.domain.crypto.key.WalletDto
 import org.springframework.stereotype.Service
 
 @Service
@@ -15,7 +17,8 @@ class DefaultCryptoService(
     private val seedCalculator: SeedCalculator,
     private val derivationKeyHelper: DerivationKeysHelper,
     private val extendedKeySerializer: ExtendedKeySerializer,
-    private val seedPhraseValidator: SeedPhraseValidator
+    private val seedPhraseValidator: SeedPhraseValidator,
+    private val derivationKeysHelper: DerivationKeysHelper
 ) : CryptoService {
 
     override fun generateSeedPhrase(): String = seedPhraseGenerator.createSeedPhrase(TWELVE)
@@ -37,5 +40,17 @@ class DefaultCryptoService(
     override fun serializedPublicKey(key: ExtendedKey): String = extendedKeySerializer.serializePublic(key)
 
     override fun serializedPrivateKey(key: ExtendedKey): String = extendedKeySerializer.serializePrivate(key)
+
+    override fun generateKey(): WalletDto {
+        val seedPhrase = seedPhraseGenerator.createSeedPhrase(TWELVE)
+        val rootExtendedKey = ExtendedKey.root(seedPhrase.toByteArray())
+
+        val extendedKey = derivationKeysHelper.deriveDefaultAddress(rootExtendedKey)
+        val addressKeyDto = AddressKeyDto(extendedKeySerializer.serializePublic(extendedKey),
+                extendedKeySerializer.serializePrivate(extendedKey), extendedKey.ecKey.getAddress())
+
+        return WalletDto(seedPhrase, extendedKeySerializer.serializePublic(rootExtendedKey),
+                extendedKeySerializer.serializePrivate(rootExtendedKey), addressKeyDto)
+    }
 
 }