+++ src/test/kotlin/io/openfuture/chain/service/CryptoServiceTests.kt
@@ -7,6 +7,7 @@ import io.openfuture.chain.crypto.key.DerivationKeysHelper
 import io.openfuture.chain.crypto.key.ExtendedKeySerializer
 import io.openfuture.chain.crypto.seed.PhraseLength
 import io.openfuture.chain.crypto.seed.calculator.SeedCalculator
+import io.openfuture.chain.crypto.seed.PhraseLength.TWELVE
 import io.openfuture.chain.crypto.seed.generator.SeedPhraseGenerator
 import io.openfuture.chain.crypto.seed.validator.SeedPhraseValidator
 import org.assertj.core.api.Assertions.assertThat
@@ -18,6 +19,8 @@ import org.mockito.Mock
 class CryptoServiceTests : ServiceTests() {
 
     @Mock private lateinit var seedPhraseGenerator: SeedPhraseGenerator
+
+    @Mock private lateinit var derivationKeysHelper: DerivationKeysHelper
 
     @Mock private lateinit var seedCalculator: SeedCalculator
 
@@ -45,7 +48,7 @@ class CryptoServiceTests : ServiceTests() {
     fun generateSeedPhraseShouldReturnTwelveSeedPhraseWords() {
         val seedPhrase = "1 2 3 4 5 6 7 8 9 10 11 12"
 
-        given(seedPhraseGenerator.createSeedPhrase(PhraseLength.TWELVE)).willReturn(seedPhrase)
+        given(seedPhraseGenerator.createSeedPhrase(TWELVE)).willReturn(seedPhrase)
 
         val seedPhraseResult = cryptoService.generateSeedPhrase()
 
@@ -112,6 +115,24 @@ class CryptoServiceTests : ServiceTests() {
         assertThat(key.ecKey.private).isNotNull()
         assertThat(key.parentFingerprint).isNotNull()
         assertThat(key.sequence).isNotNull()
+    fun generateKeyShouldReturnWalletDtoTest() {
+        val seedPhrase = "ability able about above absent absorb abstract absurd abuse accident access act"
+        val privateKey  = "xprv9s21ZrQH143K4QKw9Cq9BUSUJGMSNMBt5mQVU8QD32NZpw4i6bnmiACNkqunc6P6B5tHXGw4oJMo2wXVwDgj2WDQFpTFufd4TdtKpZvpgEb"
+        val publicKey = "xpub661MyMwAqRbcGtQQFEN9YcPCrJBvmoujSzL6GWopbMuYhjPre972FxWrc6NHZiH87hAz3vg3o95GDTwncHF6dMkoJLQ897p4VssRDA4kJ7V"
+
+        given(seedPhraseGenerator.createSeedPhrase(TWELVE)).willReturn(seedPhrase)
+        given(derivationKeysHelper.deriveDefaultAddress(any(ExtendedKey::class.java))).will { invocation -> invocation.arguments[0]  }
+        given(extendedKeySerializer.serializePublic(any(ExtendedKey::class.java))).willReturn(publicKey)
+        given(extendedKeySerializer.serializePrivate(any(ExtendedKey::class.java))).willReturn(privateKey)
+
+        val actualWalletDto = cryptoService.generateKey()
+
+        assertThat(actualWalletDto.seedPhrase).isEqualTo(seedPhrase)
+        assertThat(actualWalletDto.masterPublicKey).isEqualTo(publicKey)
+        assertThat(actualWalletDto.masterPrivateKey).isEqualTo(privateKey)
+        assertThat(actualWalletDto.addressKeyDto.publicKey).isEqualTo(publicKey)
+        assertThat(actualWalletDto.addressKeyDto.privateKey).isEqualTo(privateKey)
+        assertThat(actualWalletDto.addressKeyDto.address).isNotNull()
     }
 
 }